requirements.txt
# Core
streamlit==1.39.0
python-dotenv==1.0.1

# PII Detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.6

# NER Training
datasets==2.21.0
huggingface-hub==0.24.6

# Vector Store & Embeddings
chromadb==0.5.5
sentence-transformers==3.0.1

# Encryption
cryptography==43.0.1

# LLM (Groq API)
httpx==0.27.2

# Experiment Tracking
mlflow==2.16.2

# Document Processing
PyPDF2==3.0.1

# Data & Visualization
numpy==1.26.4
plotly==5.24.1
pandas==2.2.3

# Testing
pytest==8.3.3
